requirements.txt
pandas
GoogleNews
tqdm
newspaper3k